dev last

+++ src/service/sys_trash_service.rs
@@ -66,7 +66,7 @@ impl SysTrashService {
     //recycle trash older than `trash_recycle_days`
     pub async fn recycle(&self) -> Result<u64, Error> {
         let before = DateTime::now().0.sub(Duration::from_secs(
-            CONTEXT.config.trash_recycle_days as u64 * 24 * 3600,
+            CONTEXT.config.trash_recycle_days * 24 * 3600,
         ));
         let r = SysTrash::delete_by_day_before(pool!(), DateTime(before)).await?;
         Ok(r.rows_affected)
